Very simple flask template for data ingestion.
This is an ultra-simple template that allows you to create a database,
run a web app (using flask), and then send some data to that web app.
The web app then uses the SQLAlchemy ORM to save the data to an sqlite
database.

THIS HAS NO ASSERTS. It does not check that the received data actually
matches the shape of the database. It's purely here as a very-basic
barebones kickstart to get a server saving data.